# Conflicts:
#	pom.xml
#	src/main/java/ru/zagshak/buySupply/domain/Offer.java
#	src/main/java/ru/zagshak/buySupply/service/OfferService.java

+++ src/main/java/ru/zagshak/buySupply/service/EstimateService.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 import ru.zagshak.buySupply.domain.Estimate;
 import ru.zagshak.buySupply.repository.estimateRepo.EstimateRepoImpl;
+import ru.zagshak.buySupply.util.exception.NoAccessException;
+
 import static ru.zagshak.buySupply.util.ValidationUtil.checkNotFoundWithId;
 
 
@@ -20,6 +22,9 @@ public class EstimateService {
 
     public void update(Estimate estimate, int estimatedId, int estimatorId) {
         Assert.notNull(estimate, "estimate must not be null");
+        if (estimate.getEstimator().getId() != estimatorId) {
+            throw new NoAccessException("You can't change others estimates");
+        }
         checkNotFoundWithId(repo.save(estimate, estimatedId,estimatorId), estimate.getId());
     }
 
@@ -32,19 +37,23 @@ public class EstimateService {
 
 
     public boolean delete(int id,int estimatorId) {
+        if (get(id).getEstimator().getId() != estimatorId) {
+            throw new NoAccessException("You can't change others estimates");
+        }
         return repo.delete(id,estimatorId);
     }
 
 
     public Estimate get(int id) {
-        return repo.get(id);
+        return checkNotFoundWithId(repo.get(id), id);
     }
 
+
+    public List<Estimate> getAll() { return repo.getAll(); }
 
     public List<Estimate> getAllByEstimator(int estimatorId) {
         return repo.getAllByEstimator(estimatorId);
     }
-
 
     public List<Estimate> getAllByEstimated(int estimatedId) {
         return repo.getAllByEstimated(estimatedId);
